+++ src/languages/QOwnNotes_en.ts
@@ -4350,12 +4350,12 @@ If the trash is enabled on your ownCloud server you should be able to restore th
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../widgets/notepreviewwidget.cpp" line="323"/>
+        <location filename="../widgets/notepreviewwidget.cpp" line="365"/>
         <source>HTML files</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../widgets/notepreviewwidget.cpp" line="324"/>
+        <location filename="../widgets/notepreviewwidget.cpp" line="366"/>
         <source>Export preview as raw HTML file</source>
         <translation type="unfinished"></translation>
     </message>
@@ -4577,17 +4577,17 @@ If the trash is enabled on your ownCloud server you should be able to restore th
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../entities/note.cpp" line="2600"/>
+        <location filename="../entities/note.cpp" line="2640"/>
         <source>Could not read image file: %1</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../entities/note.cpp" line="3221"/>
+        <location filename="../entities/note.cpp" line="3261"/>
         <source>Note file path changed</source>
         <translation type="unfinished"></translation>
     </message>
     <message numerus="yes">
-        <location filename="../entities/note.cpp" line="3222"/>
+        <location filename="../entities/note.cpp" line="3262"/>
         <source>A change of the note path was detected. Would you like to replace all occurrences of &lt;strong&gt;%1&lt;/strong&gt; links with &lt;strong&gt;%2&lt;/strong&gt; and links with filename &lt;strong&gt;%3&lt;/strong&gt; with &lt;strong&gt;%4&lt;/strong&gt; in &lt;strong&gt;%n&lt;/strong&gt; note file(s)?</source>
         <translation type="unfinished">
             <numerusform></numerusform>
@@ -4595,7 +4595,7 @@ If the trash is enabled on your ownCloud server you should be able to restore th
         </translation>
     </message>
     <message numerus="yes">
-        <location filename="../entities/note.cpp" line="3802"/>
+        <location filename="../entities/note.cpp" line="3842"/>
         <source>…and %n more note(s)</source>
         <translation type="unfinished">
             <numerusform></numerusform>
@@ -4624,22 +4624,22 @@ If the trash is enabled on your ownCloud server you should be able to restore th
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../utils/gui.cpp" line="977"/>
+        <location filename="../utils/gui.cpp" line="988"/>
         <source>Double-click to unstick note from tab</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../utils/gui.cpp" line="978"/>
+        <location filename="../utils/gui.cpp" line="989"/>
         <source>Double-click to stick note to tab</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../utils/gui.cpp" line="1020"/>
+        <location filename="../utils/gui.cpp" line="1031"/>
         <source>&lt;strong&gt;%1&lt;/strong&gt;&lt;br /&gt;last modified: %2</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../utils/gui.cpp" line="1025"/>
+        <location filename="../utils/gui.cpp" line="1036"/>
         <source>&lt;br /&gt;path: %1</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8400,6 +8400,7 @@ Your notes will stay intact!</source>
     </message>
     <message>
         <location filename="../dialogs/tododialog.ui" line="157"/>
+        <location filename="../dialogs/tododialog.ui" line="326"/>
         <source>Summary</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8411,6 +8412,16 @@ Your notes will stay intact!</source>
     <message>
         <location filename="../dialogs/tododialog.ui" line="222"/>
         <source>Note…</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../dialogs/tododialog.ui" line="346"/>
+        <source>This doesn&apos;t work for sub-items, because they may be hidden by the parent item!</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../dialogs/tododialog.ui" line="349"/>
+        <source>Show only items due today</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -8444,7 +8455,12 @@ Your notes will stay intact!</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../dialogs/tododialog.ui" line="334"/>
+        <location filename="../dialogs/tododialog.ui" line="331"/>
+        <source>Due date</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../dialogs/tododialog.ui" line="339"/>
         <source>select your todo list</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8499,33 +8515,33 @@ Your notes will stay intact!</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../dialogs/tododialog.cpp" line="582"/>
+        <location filename="../dialogs/tododialog.cpp" line="606"/>
         <source>Remove todo item</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../dialogs/tododialog.cpp" line="583"/>
+        <location filename="../dialogs/tododialog.cpp" line="607"/>
         <source>Remove the selected todo item?
 This cannot be undone!</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../dialogs/tododialog.cpp" line="867"/>
+        <location filename="../dialogs/tododialog.cpp" line="891"/>
         <source>Create sub-task</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../dialogs/tododialog.cpp" line="874"/>
+        <location filename="../dialogs/tododialog.cpp" line="898"/>
         <source>Create new sub-task</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../dialogs/tododialog.cpp" line="874"/>
+        <location filename="../dialogs/tododialog.cpp" line="898"/>
         <source>Name:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../dialogs/tododialog.cpp" line="875"/>
+        <location filename="../dialogs/tododialog.cpp" line="899"/>
         <source>New sub-task</source>
         <translation type="unfinished"></translation>
     </message>